Snake changed
Changed snake, deleted unnecessary file
