+++ .idea/compiler.xml
@@ -6,8 +6,12 @@
         <sourceOutputDir name="target/generated-sources/annotations" />
         <sourceTestOutputDir name="target/generated-test-sources/test-annotations" />
         <outputRelativeToContentRoot value="true" />
+        <module name="snake" />
         <module name="snake_game" />
       </profile>
     </annotationProcessing>
+    <bytecodeTargetLevel>
+      <module name="snake" target="6" />
+    </bytecodeTargetLevel>
   </component>
 </project>
