+++ src/main/java/sample/src/sample/Snake.java
@@ -5,33 +5,38 @@ import javafx.scene.Group;
 import javafx.scene.paint.Color;
 import javafx.scene.shape.Rectangle;
 import javafx.stage.Stage;
-import org.apache.log4j.Logger;
 
 import java.util.LinkedList;
 
 public class Snake {
 
-    private static long frameDelay = 250000000; 
-    private Rectangle head = new Rectangle(20, 20); 
+
+	private static final long START_FRAMEDELAY = 250000000;
+	private static final long MAX_FRAMEDELAY = 80000;
+	private static final long DECREASE_DELAY = 6000000;  //von speedRefresh abziehen
+    public long frameDelay = START_FRAMEDELAY; //250-300 mill. guter Startwert
+    private int helpX;
+    private int helpY;
+    private static int rectangleWidth = 20;
+    private static int rectangleHeight = 20;
+
+    //GameObject food = new GameObject();
+    private Rectangle head = new Rectangle(20, 20); // hier Initialisiert, weil in mehreren Methoden
     private LinkedList<Rectangle> snake = new LinkedList<Rectangle>();
-    
 
-
-    private static org.apache.log4j.Logger logSnake = org.apache.log4j.Logger.getLogger(Snake.class);
-
-    Snake(Group group, Stage stage) {
+    public Snake(Group group, Stage stage) {
         snake.add(head);
         snake.getFirst().relocate(stage.getWidth() / 2, stage.getHeight() / 2);
         group.getChildren().add(snake.getFirst());
 
     }
 
-    long getframeDelay() {
+    public long getframeDelay() {
         return frameDelay;
     }
 
 
-    void respawn(Group group, GameObject food, Score score, Stage stage, Control control) {
+    public void respawn(Group group, GameObject food, Score score, Stage stage, Control control) {
         group.getChildren().clear();
         snake.clear();
 
@@ -39,83 +44,78 @@ public class Snake {
         snake.getFirst().relocate(stage.getWidth() / 2, stage.getHeight() / 2);
         group.getChildren().add(snake.getFirst());
         food.setFood(group, stage); // setet neues random food und getchilded es
-        score.scoreRespawn(group); // respawn Mehtode fÃ¼r Score
-        logSnake.info("respawn");
-        logSnake.debug("respawn");
-        frameDelay = 250000000; // zurÃ¼ck zum Standardwert
-        logSnake.debug("back to value: "+frameDelay);
+        score.scoreRespawn(group); // respawn Mehtode für Score
+        frameDelay = START_FRAMEDELAY; // zurück zum Standardwert
+
+
         control.stopMovement();
 
     }
 
-    void snakeDead(Group group, Control control, Stage stage) {
-        //Last Minute - wird gebraucht um Score nicht zu frÃ¼h zu lÃ¶schen (Ã¼berlegung nur respawn zu verwenden mit dieser implementierung fehlgeschlagen)
-
-        logSnake.info("snake dead!");
-        logSnake.debug("snake dead!");
+    public void snakeDead(Group group, Control control, Stage stage) {
+        //Last Minute - wird gebraucht um Score nicht zu früh zu löschen (überlegung nur respawn zu verwenden mit dieser implementierung fehlgeschlagen)
 
         group.getChildren().clear();
         snake.clear();
         snake.add(head);
         snake.getFirst().relocate(stage.getWidth() / 2, stage.getHeight() / 2);
-        frameDelay = 250; // zurÃ¼ck zum Standardwert
+
+        frameDelay = START_FRAMEDELAY; // zurück zum Standardwert
         control.stopMovement();
 
     }
 
 
-    private void eat(Group group, Score score, GameObject food) {//added ein tail rectangle, Ã¼bernimmt color von food,erhÃ¶ht score um 1, macht schneller
+    private void eat(Group group, Score score, GameObject food) {//added ein tail rectangle, übernimmt color von food,erhöht score um 1, macht schneller
         snake.add(new Rectangle(20, 20));
-        logSnake.debug("eat");
-        snake.getLast().setFill(Color.color(food.getColor()[0], food.getColor()[1], food.getColor()[2])); //holt sich aus deathsoundMedia GameObject die Color von Food fÃ¼r sein Tail
+        snake.getLast().setFill(Color.color(food.getColor()[0], food.getColor()[1], food.getColor()[2])); //holt sich aus deathsoundMedia GameObject die Color von Food für sein Tail
         group.getChildren().add(snake.getLast()); //bringt den tail auf die Szene
         score.upScoreValue(); // added +1 zu scoreValue
-        if (frameDelay >= 80000) { //maximale Grenze sonst wirds zu schnell
-            //von speedRefresh abziehen
-            long delayDecrease = 600000;
-            frameDelay -= delayDecrease;
+
+        if (frameDelay >= MAX_FRAMEDELAY) { //maximale Grenze sonst wirds zu schnell
+            frameDelay -= DECREASE_DELAY;
             System.out.println(frameDelay);
         }
 
     }
 
-    public void collision(GameObject food, Group group, Bounds foodBound, Score score, Control control, Stage stage, Gameboard gameboard) { //gameobject sind obstacles so wie Food, Boundarys fÃ¼r Collisions
+    public void collision(GameObject food, Group group, Bounds foodBound, Score score, Control control, Stage stage, Gameboard gameboard) { //gameobject sind obstacles so wie Food, Boundarys für Collisions
         Bounds headBox = head.getBoundsInParent(); // erstellt eine Boundary um den Snakekopf
 
 
-
-        if (headBox.intersects(foodBound)) {
-        	eat(group, score, food);
+        if (headBox.intersects(foodBound)) {//überprüfung Collision Head mit Food Boundary
+            eat(group, score, food);
             food.setFood(group, stage);
             GameLoop.playEatsound();
         }
 
-        if (head.getLayoutX() <= 0 || head.getLayoutX() >= stage.getWidth() - 30 || // ÃœberprÃ¼fung ob Head den Rand trifft
+        if (head.getLayoutX() <= 0 || head.getLayoutX() >= stage.getWidth() - 30 || // Überprüfung ob Head den Rand trifft
                 head.getLayoutY() <= 0 || head.getLayoutY() >= stage.getHeight() - 54) {
             snakeDead(group, control, stage);
             gameboard.setDeathTouchWall(score);
             GameLoop.playDeathsound();
-            GameLoop.stopIngamemusic();
-            GameLoop.restartGameovermusic();
+            //GameLoop.stopIngamemusic();
+            //GameLoop.restartGameovermusic();
         }
 
 
-        for (int i = 1; i < this.snake.size(); i++) { //ÃœberprÃ¼fung Snake beisst sich in den oasch
+        for (int i = 1; i < this.snake.size(); i++) { //Überprüfung Snake beisst sich in den oasch
             if (headBox.intersects(this.snake.get(i).getBoundsInParent())) {
                 System.err.println("Dead");
                 snakeDead(group, control, stage);
                 gameboard.setDeathTouchTail(score);
                 GameLoop.playDeathsound();
-                GameLoop.stopIngamemusic();
-                GameLoop.restartGameovermusic();
+                //GameLoop.stopIngamemusic();
+                //GameLoop.restartGameovermusic();
             }
         }
     }
 
 
     public void moveSnake(int dx, int dy, Stage stage) { //dx bzw dy ist jeweils + oder - speed, war zuvor 5
-
-        if (dx != 0 || dy != 0) { //gibt es Ã¼berhaupt dx/dy werte (wenn wir stehen z.B. nicht)
+        int helpX;
+        int helpY;
+        if (dx != 0 || dy != 0) { //gibt es überhaupt dx/dy werte (wenn wir stehen z.B. nicht)
             LinkedList<Rectangle> snakehelp = new LinkedList<Rectangle>();
 
             for (int i = 0; i < snake.size(); i++) {
@@ -125,15 +125,13 @@ public class Snake {
                 snakehelp.get(i).relocate(snake.get(i).getLayoutX(), snake.get(i).getLayoutY());
             }
 
-            int x = (int) snake.getFirst().getLayoutX() + dx;
-            int y = (int) snake.getFirst().getLayoutY() + dy;
-            snake.getFirst().relocate(x, y);//moved erstmal nur den Kopf
+            snake.getFirst().relocate((int) snake.getFirst().getLayoutX() + dx, (int) snake.getFirst().getLayoutY() + dy);//moved erstmal nur den Kopf
 
 
             for (int i = 1; i < snake.size(); i++) {
 
-                int helpX = (int) snakehelp.get(i - 1).getLayoutX();
-                int helpY = (int) snakehelp.get(i - 1).getLayoutY();
+                helpX = (int) snakehelp.get(i - 1).getLayoutX();
+                helpY = (int) snakehelp.get(i - 1).getLayoutY();
                 snake.get(i).relocate(helpX, helpY);
             }
         }
